requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
argon2-cffi==23.1.0
async-timeout==5.0.1
attrs==26.1.0
autobahn==19.11.2
Automat==25.4.16
bcrypt==4.2.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.6
exceptiongroup==1.3.1
fastapi==0.135.2
h11==0.16.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiteconnect==5.1.0
MarkupSafe==3.0.3
packaging==26.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyOpenSSL==26.0.0
pyotp==2.9.0
python-dateutil==2.9.0.post0
python-multipart==0.0.22
pywin32==311
redis==7.4.0
requests==2.33.1
service-identity==24.2.0
six==1.17.0
starlette==1.0.0
tomli==2.4.1
Twisted==25.5.0
txaio==25.9.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
zope.interface==8.2
qrcode==7.4.2
Pillow==10.4.0
